summary: Separation of libasm compilation and build process
It is now possible to disable some part of libasm
deepending on needs.

With the new flags --libasm-ia32, --libasm-sparc, --libasm-mips
of the configure script, it is possible

Another changed as been made:
Makefile is not generated anymore.
Some rules may be missing to compile kernsh, e2dbg or
some other part.

Todo list:

A list of the different targets possible should be done :

with kernsh, without
32 or 64 bits
with or without readline
with libasm arch support.