Trying to dettach the network code from the rest of the code.
We try to definitely avoid the use of ERESI_NET, it brakes the code,
make it really hard to read and hard to maintain. Instead, we declare
a function revm_is_net_supported() returning true if the networking
features are turned on (by the compilation). The drawback of that is
revm_prepare_select has some embedded code in it. We should absolutely
separate the net part from the rest in revm_prepare_select.

I introduced a new formating for function description , I like it :P,
but of course if it doesn't work with doxygen or you don't like it,
I'll reverse it. I'll probably remove it actually, since the best it
to keep regularity through the program.

A couple of function need to be private to the file and not be
available outside this scope. This is true at least for the networking
stuff. For example, "revm_check_net_select" is used only in select.c
and thus, must be "private" to this file.

I introduced a segfault in rcmd btw. But the function needs to be
rewritten, so I'll work on that later on.